Bug 1407707 — Add conditional edges to the screen graph.
This lets us define edges (including actions) as conditional on the current userstate.

If the userState changes (even mid `goto`), the conditional edges are re-evaluated, and the navigator is re-routed.

Unfortunately, this commit adds mutable state to the graph, which means that there is now a one-to-one navigator-to-screengraph mapping (before you could have multiple navigators from the same instance of the ScreenGraph. This is almost certainly not worth worrying about.